Range check winsock2's socket() return value
Winsock2's socket() returns the unsigned type SOCKET, which is a 32-bit
type for WIN32 and a 64-bit type for WIN64.

It seems we can always safely assign SOCKET to an int: failure is
indicated by INVALID_SOCKET which will cast to -1 as an int; when SOCKET
is a 32-bit type other values will cast appropriately; when SOCKET is a
64-bit type it seems in practice that valid values all fit in 32-bits
(and that we're not the only code to assume this, so Microsoft are
unlikely to arbitrarily change that).

But we should check and throw an exception rather than quietly mangling
the value.