Stop using _set_errno()
This is a Microsoft-specific function for setting errno, which we
were using in just one place in the code - in more than 30 other
places in core we just assign to errno directly.

The docs don't give a reason why _set_errno() is needed, and at
least the implementation in Wine is exactly equivalent to just
assigning to errno.