Address GCC13 -Wredunant-move warning
This std::move() was added in e23164b4ddb6cdc57136f4da2598d831f3780035
to fix a clang warning from -Wreturn-std-move (enabled by -Wall).  Just
removing the std::move() reintroduces that warning with clang 8 and
clang 11 (but not clang 13 or later) but adding a static cast to the
returned type seems to make all versions of both compilers happy.